ndarray major change: use a class as shape parameter
This unlocks a critical optimization: using partially constant shapes so
that some loop may be unrolled by the compiler.

Unfortunately, changing the ndarray type (and the associated shapes for
the various views of ndarray) as a large impact on the codebase, but
that's for the greater good, isn't it?

+++ pythran/pythonic/utils/nested_container.hpp
@@ -12,7 +12,7 @@ namespace utils
 {
 
   template <class T>
-  size_t nested_container_size<T>::flat_size(T const &t)
+  long nested_container_size<T>::flat_size(T const &t)
   {
     return t.size() *
            nested_container_size<typename std::conditional<
@@ -24,7 +24,7 @@ namespace utils
 
   /* Recursion stops on bool */
   template <class F>
-  constexpr size_t nested_container_size<bool>::flat_size(F)
+  constexpr long nested_container_size<bool>::flat_size(F)
   {
     return 1;
   }
